+++ golem/cli/argument_parser.py
@@ -11,18 +11,16 @@ def get_parser():
     parser_run = subparsers.add_parser('run', add_help=False)
     parser_run.add_argument('project', nargs='?', default='')
     parser_run.add_argument('test_query', nargs='?', default='')
-    parser_run.add_argument('-b', '--browsers', action='store',
-                            nargs='*', default=[], type=str)
-    parser_run.add_argument('-t', '--threads', action='store', nargs='?',
-                            default=1, type=int)
-    parser_run.add_argument('-e', '--environments', action='store',
-                            nargs='+', default=[], type=str)
+    parser_run.add_argument('-b', '--browsers', nargs='*', default=[], type=str)
+    parser_run.add_argument('-p', '--processes', nargs='?', default=1, type=int)
+    parser_run.add_argument('-e', '--environments', nargs='+', default=[], type=str)
+    parser_run.add_argument('-t', '--tags', nargs='*', default=[], type=str)
     parser_run.add_argument('-i', '--interactive', action='store_true', default=False)
-    parser_run.add_argument('-r', '--report', action='store', nargs='+',
-                            choices=report_choices, default=[], type=str)
-    parser_run.add_argument('--report-folder', action='store', nargs='?', type=str)
-    parser_run.add_argument('--report-name', action='store', nargs='?', type=str)
-    parser_run.add_argument('--timestamp', action='store', nargs='?', type=str)
+    parser_run.add_argument('-r', '--report', nargs='+', choices=report_choices,
+                            default=[], type=str)
+    parser_run.add_argument('--report-folder', nargs='?', type=str)
+    parser_run.add_argument('--report-name', nargs='?', type=str)
+    parser_run.add_argument('--timestamp', nargs='?', type=str)
     parser_run.add_argument('-h', '--help', action='store_true')
 
     # gui